+++ src/measurement/feature_min_val.h
@@ -7,8 +7,6 @@
  */
 
 
-#include <array>
-
 #include "diplib.h"
 #include "diplib/measurement.h"